arm functions: move to arm directory
