+++ src/Makefile
@@ -15,7 +15,7 @@ OBJS = main.o kernel.o minidexed.o config.o userinterface.o uimenu.o uitostring.
        zyn/AnalogFilter.o zyn/ValueSmoothingFilter.o zyn/WaveShapeSmps.o zyn/Distortion.o \
        zyn/CombFilterBank.o zyn/Sympathetic.o \
        butter.o \
-       arm_float_to_q23.o arm_zip_f32.o arm_scale_zip_f32.o \
+       arm/arm_float_to_q23.o arm/arm_zip_f32.o arm/arm_scale_zip_f32.o \
        net/ftpdaemon.o net/ftpworker.o net/applemidi.o net/udpmidi.o net/mdnspublisher.o udpmididevice.o
 
 EXTRACLEAN = $(OBJS) $(OBJS:.o=.d)
